Bug fix for JWT_TOKEN_VALIDITY

+++ microserviceBdd/microservice-web/src/main/java/com/eSales/microserviceWeb/security/JwtTokenUtil.java
@@ -17,7 +17,7 @@ import io.jsonwebtoken.SignatureAlgorithm;
 public class JwtTokenUtil implements Serializable {
 
     private static final long serialVersionUID = -2550185165626007488L;
-    public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+    public static final long JWT_TOKEN_VALIDITY = (long)5 * 60 * 60;
     @Value("badj")
     private String secret;
